% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulators.R
\name{volesWrap}
\alias{volesWrap}
\title{Wrapper for volesSimulator, used to simulated from volesSimulator() within the synlik package.}
\usage{
volesWrap(param, nsim, extraArgs, nObs, ...)
}
\description{
Wrapper for volesSimulator, used to simulated from volesSimulator() within the synlik package.
}

